pyproject.toml
[tool.poetry]
name = "adventofcode_2024"
version = "0.1.0"
description = ""
authors = ["All <allmacdev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
